Fix #91
WCSessionDelegate is also moved to AppDelegate
